+++ src/main/java/com/example/lockly/controllerLayer/dtos/FolderDto.java
@@ -3,4 +3,4 @@ package com.example.lockly.controllerLayer.dtos;
 import lombok.Builder;
 
 @Builder
-public record FolderDto(Long id, String name, UserDto userDto){}
+public record FolderDto(Integer id, String name, UserDto userDto){}
