+++ src/main/java/com/example/lockly/controllerLayer/dtos/UserDto.java
@@ -3,4 +3,4 @@ package com.example.lockly.controllerLayer.dtos;
 import lombok.Builder;
 
 @Builder
-public record UserDto(Long id, String name, String email, String password) {}
+public record UserDto(Integer id, String name, String email, String password) {}
