+++ src/main/java/com/example/lockly/repositoryLayer/FolderRepository.java
@@ -1,9 +1,10 @@
 package com.example.lockly.repositoryLayer;
 
 import com.example.lockly.repositoryLayer.entities.FolderEntity;
+import com.example.lockly.repositoryLayer.entities.pfk.FolderId;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FolderRepository extends JpaRepository<FolderEntity, Long> {
+public interface FolderRepository extends JpaRepository<FolderEntity, Integer> {
 }
