+++ src/main/java/com/example/lockly/serviceLayer/passwords/PasswordWithoutFolderService.java
@@ -3,8 +3,8 @@ package com.example.lockly.serviceLayer.passwords;
 import com.example.lockly.dataproviderLayer.PasswordWithoutFolderDataProvider;
 import com.example.lockly.domainLayer.passwords.PasswordWithoutFolder;
 import com.example.lockly.serviceLayer.UserService;
-import com.example.lockly.serviceLayer.exceptions.NoPasswordFoundException;
-import com.example.lockly.serviceLayer.exceptions.PasswordNotFoundException;
+import com.example.lockly.serviceLayer.exceptions.password.NoPasswordFoundException;
+import com.example.lockly.serviceLayer.exceptions.password.PasswordNotFoundException;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -25,7 +25,7 @@ public class PasswordWithoutFolderService {
         return dataProvider.save(newPassword);
     }
 
-    public List<PasswordWithoutFolder> consultAllByUser(Long idUser) {
+    public List<PasswordWithoutFolder> consultAllByUser(Integer idUser) {
         List<PasswordWithoutFolder> result = dataProvider.consultAllByUser(idUser);
         if(result.isEmpty()) {
             throw new NoPasswordFoundException();
@@ -41,12 +41,12 @@ public class PasswordWithoutFolderService {
         return result.get();
     }
 
-    public void delete(Long id) {
+    public void delete(Integer id) {
         consultById(id);
         dataProvider.delete(id);
     }
 
-    public PasswordWithoutFolder change(PasswordWithoutFolder newData, Long id) {
+    public PasswordWithoutFolder change(PasswordWithoutFolder newData, Integer id) {
         PasswordWithoutFolder existingPassword = consultById(id);
         if(newData.getUser() != null)
             existingPassword.setUser(newData.getUser());
@@ -54,7 +54,7 @@ public class PasswordWithoutFolderService {
         return dataProvider.save(existingPassword);
     }
 
-    public PasswordWithoutFolder consultById(Long idPasswordWithoutFolder) {
+    public PasswordWithoutFolder consultById(Integer idPasswordWithoutFolder) {
         Optional<PasswordWithoutFolder> result = dataProvider.consultById(idPasswordWithoutFolder);
         if(result.isEmpty())
             throw new PasswordNotFoundException();
